Web_Scraping/Pinterest_Scraper/doesnt_work/requirements.txt




requests
beautifulsoup4
PyYAML
